Press2Emoji/requirements.txt
Pillow
keyboard
